pyproject.toml
[project]
name = "mobile-glm"
version = "0.1.0"
description = "Phone automation agent using Claude Agent SDK + AutoGLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.18",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "websockets>=12.0",
    "adbutils>=2.0.0",
    "numpy>=1.24.0",
    "phone-agent @ git+https://github.com/zai-org/Open-AutoGLM.git",
]

[tool.setuptools]
py-modules = ["agent_sdk", "phone_tool", "preference_tool", "scrcpy_ws_bridge"]
